Fix context selection to meet current design

+++ src/datasets/context_dataset.py
@@ -31,5 +31,5 @@ class ContextDataset(BaseDataset):
 
         return {
             "input_ids": torch.as_tensor(input_ids, dtype=torch.long),
-            "label": int(d.get("has_answer")),
+            "label": float(d.get("has_answer")),
         }
